requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
pip==25.1.1
pyperclip==1.9.0
pyshorteners==1.0.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
